### 2024-03-21 #6518 @editorialbot
[PRE REVIEW]: GeophysicalModelGenerator.jl: A Julia package to visualise geoscientific data and create numerical model setups
**Submitting author:** @boriskaus (<a href="http://orcid.org/0000-0002-0247-8660">Boris Kaus</a>)
**Repository:** https://github.com/JuliaGeodynamics/GeophysicalModelGenerator.jl
**Branch with paper.md** (empty if default branch): bk-joss-paper
**Version:** v0.7.0
