+++ src/utils.jl
@@ -79,6 +79,20 @@ function addfield(V::Q1Data,new_fields::NamedTuple; cellfield=false)
 end
 
 
+"""
+    V = addfield(V::FEData,new_fields::NamedTuple; cellfield=false)
+
+Add `new_fields` fields to a `FEData` dataset; set `cellfield` to `true` if the field is a cell field; otherwise it is a vertex field
+"""
+function addfield(V::FEData,new_fields::NamedTuple; cellfield=false)
+    if cellfield
+        return FEData(V.vertices, V.connectivity, V.fields, merge(V.cellfields, new_fields))
+    else
+        return FEData(V.vertices, V.connectivity, merge(V.fields, new_fields), V.cellfields)
+    end
+end
+
+
 # this function is taken from @JeffreySarnoff
 function dropnames(namedtuple::NamedTuple, names::Tuple{Vararg{Symbol}})
     keepnames = Base.diff_names(Base._nt_names(namedtuple), names)
