+++ test/runtests.jl
@@ -33,6 +33,10 @@ end
     include("test_lamem.jl")
 end
 
+@testset "pTatin" begin
+    include("test_pTatin_IO.jl")
+end
+
 @testset "SetupGeometry" begin
     include("test_setup_geometry.jl")
 end
