+++ src/GeophysicalModelGenerator.jl
@@ -39,6 +39,7 @@ include("Paraview_collection.jl")
 include("transformation.jl")
 include("voxel_gravity.jl")
 include("LaMEM_io.jl")
+include("pTatin_IO.jl")
 include("Setup_geometry.jl")
 include("stl.jl")
 include("ProfileProcessing.jl")
