### 2024-03-29 fc25a7a
make Q1Data work with add_box! and friends
