### 2024-03-29 759c901
add extrema for various data types
