+++ src/GeophysicalModelGenerator.jl
@@ -71,10 +71,18 @@ export import_topo, import_GeoTIFF
         visualise
 Interactive widget that allows you to explore a 3D data set `DataSet` in an interactive manner.
 It requires you to load `GLMakie`.
-"""
+""";
 function visualise end
 export visualise
 
+
+"""
+    import_Gmsh(fname::String)
+
+Reads a Gmsh file. Requires loading `GridapGmsh`.
+"""
+function import_Gmsh end
+export import_Gmsh
 
 
 end
