+++ Project.toml
@@ -13,6 +13,7 @@ GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
 Geodesy = "0ef565a4-170c-5f04-8de2-149903a85f3d"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
+GridapGmsh = "3025c34a-b394-11e9-2a55-3fee550c04c8"
 ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
@@ -31,10 +32,12 @@ WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 [weakdeps]
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GMT = "5752ebe1-31b9-557e-87aa-f909b540aa54"
+GridapGmsh = "3025c34a-b394-11e9-2a55-3fee550c04c8"
 
 [extensions]
 GLMakie_Visualisation = "GLMakie"
 GMT_utils = "GMT"
+Gmsh_utils = "GridapGmsh"
 
 [compat]
 Colors = "0.9 - 0.12"
@@ -47,10 +50,11 @@ GeoParams = "0.2 - 0.5"
 Geodesy = "1"
 GeometryBasics = "0.1 - 0.4"
 Glob = "1.2 - 1.3"
+GridapGmsh = "0.5 - 0.7"
 ImageIO = "0.1 - 0.6"
 Interpolations = "0.14, 0.15"
-LightXML = "0.8, 0.9"
 JLD2 = "0.4"
+LightXML = "0.8, 0.9"
 MeshIO = "0.1 - 0.4"
 NearestNeighbors = "0.2 - 0.4"
 Parameters = "0.9 - 0.12"
