+++ Project.toml
@@ -13,7 +13,6 @@ GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
 Geodesy = "0ef565a4-170c-5f04-8de2-149903a85f3d"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
-GridapGmsh = "3025c34a-b394-11e9-2a55-3fee550c04c8"
 ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
@@ -70,4 +69,4 @@ StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "GMT", "StableRNGs"]
+test = ["Test", "GMT", "StableRNGs","GridapGmsh"]
