+++ .github/workflows/SpellCheck.yml
@@ -10,6 +10,6 @@ jobs:
       - name: Checkout Actions Repository
         uses: actions/checkout@v4
       - name: Check spelling
-        uses: crate-ci/typos@v1.18.2
+        uses: crate-ci/typos@v1.21.0
         with:
             args: --exclude **/*.txt --exclude **/*.pvsm
