### 2024-05-06 #6518 @rwegener2
👋🏻 @daniellivingston & @anowacki, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
