### 2024-05-14 #6518 @rwegener2
👋🏻 @AnantHariharan1996, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
