+++ Project.toml
@@ -45,7 +45,7 @@ FFMPEG = "0.4"
 FileIO = "1"
 GLMakie = "0.8, 0.9"
 GMT = "1"
-GeoParams = "0.2 - 0.5"
+GeoParams = "0.2 - 0.6"
 Geodesy = "1"
 GeometryBasics = "0.1 - 0.4"
 Glob = "1.2 - 1.3"
