### 2024-05-28 f933def
Easier handling
Changed input style  for add_polygon! Now it is the same as for add_box !
