### 2024-05-30 6c5cb22
update GLMakie dependency
