### 2024-06-04 #6763 @rwegener2
Hi @anowacki anything questions getting started with your review? The first step is to run `@editorialbot generate my checklist` to generate your checklist, then follow steps listed in the comment that gets automatically generated. Feel free to ping me if anything isn't clear!
