+++ docs/src/man/Tutorial_VolcanoModel_3D.md
@@ -1,7 +1,3 @@
-```@meta
-EditURL = "../../../tutorials/Tutorial_NumericalModel_3D.jl"
-```
-
 # Simple model of a magmatic chamber with a volcano on top 
 
 ### Aim
@@ -9,7 +5,7 @@ The aim of this tutorial is to show you how to create 3D numerical model setups 
 
 ### Generating the model
 
-Lets start with creating a 3D model setup in cartesian coordinates, which uses the `CartData` data structure, with a resolution of $ 128 \times 128 \times 128 $ inside the domain $\Omega \in [-100,100] \times [-100,100] \times [-110,50]$ km
+Lets start with creating a 3D model setup in cartesian coordinates, which uses the `CartData` data structure, with a resolution of $ 128 \times 128 \times 128 $ grid points, inside the domain $\Omega \in [-100,100] \times [-100,100] \times [-110,50]$ km
 
 ```julia
 using GeophysicalModelGenerator
@@ -57,7 +53,7 @@ add_volcano!(Phases, Temp, Grid;
     T = HalfspaceCoolingTemp(Age=20)
 )
 ```
-
+We can also add a magmatic chamber located below the volcano
 ```julia
 add_ellipsoid!(Phases, Temp, Grid; 
     cen    = (0, 0, -40), 
@@ -66,10 +62,15 @@ add_ellipsoid!(Phases, Temp, Grid;
 )
 ```
 
+where we prescribe a constant temperature of $T=1400^{\circ}C$
 ```julia
 @. Temp[Phases == 4] = 1400 
-@. Temp[Phases == 0] = 0 
 Grid = addfield(Grid, (;Phases, Temp))
+```
+
+Finally we setup the temperature of the air to $T^{\text{air}}=0^{\circ}C$
+```julia
+@. Temp[Phases == 0] = 0 
 ```
 
 ```julia
