+++ src/Setup_geometry.jl
@@ -130,11 +130,11 @@ end
 
 
 """
-    add_box!(Phase, Temp, Grid::AbstractGeneralGrid; xlim=Tuple{2}, [ylim=Tuple{2}], zlim=Tuple{2},
+    add_box!(Phase, Temp, Grid::AbstractGeneralGrid; xlim::NTuple{2, _T} = (20,100), [ylim::NTuple{2, _T} = (1,10)], zlim::NTuple{2, _T} = (10,80),
             Origin=nothing, StrikeAngle=0, DipAngle=0,
             phase = ConstantPhase(1),
             T=nothing,
-            cell=false )
+            cell=false ) where _T
 
 Adds a box with phase & temperature structure to a 3D model setup.  This simplifies creating model geometries in geodynamic models
 
@@ -189,11 +189,11 @@ julia> write_paraview(Grid,"LaMEM_ModelSetup")  # Save model to paraview
 ```
 """
 function add_box!(Phase, Temp, Grid::AbstractGeneralGrid;       # required input
-                xlim=Tuple{2}, ylim=nothing, zlim=Tuple{2},     # limits of the box
+                xlim::NTuple{2, _T} = (20,100), ylim=nothing, zlim::NTuple{2, _T} = (10,80),     # limits of the box
                 Origin=nothing, StrikeAngle=0, DipAngle=0,      # origin & dip/strike
                 phase = ConstantPhase(1),                       # Sets the phase number(s) in the box
                 T=nothing,                                      # Sets the thermal structure (various functions are available)
-                cell=false )                                    # if true, Phase and Temp are defined on cell centers 
+                cell=false ) where _T                           # if true, Phase and Temp are defined on cell centers 
 
     # Retrieve 3D data arrays for the grid
     X,Y,Z = coordinate_grids(Grid, cell=cell)
@@ -247,9 +247,10 @@ end
 
 
 """
-    add_layer!(Phase, Temp, Grid::AbstractGeneralGrid; xlim=Tuple{2}, [ylim=Tuple{2}], zlim=Tuple{2},
+    add_layer!(Phase, Temp, Grid::AbstractGeneralGrid; xlim::NTuple{2, _T} = (1,100), [ylim::NTuple{2, _T} = (0,20)], zlim::NTuple{2, _T} = (0,-100),
             phase = ConstantPhase(1),
-            T=nothing, cell=false )
+            T=nothing, cell=false ) where _T
+
 
 Adds a layer with phase & temperature structure to a 3D model setup. The most common use would be to add a lithospheric layer to a model setup.
 This simplifies creating model geometries in geodynamic models
@@ -348,9 +349,10 @@ end
 
 
 """
-    add_sphere!(Phase, Temp, Grid::AbstractGeneralGrid; cen=Tuple{3}, radius=Tuple{1},
+    add_sphere!(Phase, Temp, Grid::AbstractGeneralGrid; cen::NTuple{3, _T} = (0,0,-1), radius::NTuple{1, _T} = (0.5),
             phase = ConstantPhase(1).
-            T=nothing, cell=false )
+            T=nothing, cell=false ) where _T
+
 
 Adds a sphere with phase & temperature structure to a 3D model setup.  This simplifies creating model geometries in geodynamic models
 
@@ -390,9 +392,9 @@ julia> write_paraview(Model3D,"LaMEM_ModelSetup")           # Save model to para
 ```
 """
 function add_sphere!(Phase, Temp, Grid::AbstractGeneralGrid;    # required input
-    cen=Tuple{3}, radius=Tuple{1},                              # center and radius of the sphere
+    cen::NTuple{3, _T} = (0,0,-1), radius::NTuple{1, _T} = (0.5),    # center and radius of the sphere
[…]
