### 2024-06-19 e0a2be1
change nicer handling
Changes of the functions in the test files
