### 2024-06-28 50c827d
Bump GLMakie
