+++ Project.toml
@@ -43,7 +43,7 @@ Colors = "0.9 - 0.12"
 Downloads = "1"
 FFMPEG = "0.4"
 FileIO = "1"
-GLMakie = "0.8, 0.9, 0.10"
+GLMakie = "0.10.3"
 GMT = "1"
 GeoParams = "0.2 - 0.6"
 Geodesy = "1"
