+++ test/runtests.jl
@@ -68,6 +68,10 @@ end
     include("test_create_movie.jl")
 end
 
+@testset "Waterflow" begin
+    include("test_WaterFlow.jl")
+end
+
 # Cleanup
 foreach(rm, filter(endswith(".vts"), readdir()))
 foreach(rm, filter(endswith(".vtu"), readdir()))
