+++ Project.toml
@@ -45,7 +45,7 @@ Downloads = "1"
 FFMPEG = "0.4"
 FileIO = "1"
 GLMakie = "0.8, 0.9, 0.10"
-GMT = "1"
+GMT = "1.0 - 1.14"
 GeoParams = "0.2 - 0.6"
 Geodesy = "1"
 GeometryBasics = "0.1 - 0.4"
