### 2024-07-01 6232aba
allow specifying rainfall on the grid; also returns the area of each grid cell in m2
