+++ src/WaterFlow.jl
@@ -14,9 +14,12 @@ function spacing(lon,lat)
     dlon = zero(size(lon.val)[1:2])
     dlat = zero(size(lat.val)[1:2])
     @views dlon[2:end-1,:] = (lon.val[3:end,:,1] - lon.val[1:end-2,:,1])/2
-    @views dlon[1,:], dlon[end,:] = dlon[2,:], dlon[end-1,:]
-    @views dlat[:,2:end-1] = (lat.val[:,3:end,1] - lat.val[:,1:end-2,1])/2
-    @views dlat[:,1], dlat[:,end] = dlat[:,2], dlat[:,end-1]
+     dlon[1,:] = dlon[2,:]
+     dlon[end,:] = dlon[end-1,:]
+     dlat[:,2:end-1] = (lat.val[:,3:end,1] - lat.val[:,1:end-2,1])/2
+     dlat[:,1] = dlat[:,2]
+     dlat[:,end] = dlat[:,end-1]
+    
     return dlon, dlat
 end
 
@@ -84,7 +87,7 @@ function waterflows(Topo::GeoData, flowdir_fn= WhereTheWaterFlows.d8dir_feature;
     nin     = zeros(Int8, ni)
     c       = zeros(Int64, ni)
 
-    @views area[:,:,1], slen[:,:,1], dir[:,:,1], nout[:,:,1], nin[:,:,1], sinks, pits, c[:,:,1], bnds = waterflows(dem, cellarea, flowdir_fn;
+    area[:,:,1], slen[:,:,1], dir[:,:,1], nout[:,:,1], nin[:,:,1], sinks, pits, c[:,:,1], bnds = waterflows(dem, cellarea, flowdir_fn;
                         feedback_fn=feedback_fn, drain_pits=drain_pits, bnd_as_sink=bnd_as_sink)
 
     catchment_large =  prune_catchments(c, minsize; val=0)
