### 2024-07-03 004c293
add GLMakie heatmap and heatmap! multiple dispatch for surfaces
