### 2024-07-03 af53201
add a few more fields to simplify plotting the largest catchment
