### 2024-07-03 ac66971
allow plotting an array computed on top of a surface
