### 2024-07-03 64f4b9a
remove log10 of largest area
