+++ Project.toml
@@ -63,6 +63,7 @@ Parameters = "0.9 - 0.12"
 SpecialFunctions = "1.0, 2"
 StaticArrays = "1"
 WriteVTK = "1"
+WhereTheWaterFlows = "0.10"
 julia = "1.9"
 
 [extras]
