### 2024-07-04 0cef01a
add test for summing 2 GeoData surface grids
