### 2024-07-05 23d9d4a
add log10(largest_area)
