+++ src/Setup_geometry.jl
@@ -85,7 +85,7 @@ function add_stripes!(Phase, Grid::AbstractGeneralGrid;                # require
     cell            =  false )                          # if true, Phase and Temp are defined on cell centers 
 
     # warnings
-    if stripeWidth >= stripeSpacing/2.0
+    if stripeWidth ≥ stripeSpacing/2.0
         print("WARNING: stripeWidth should be strictly < stripeSpacing/2.0, otherwise phase is overwritten by the stripePhase\n")
     elseif sum(stripAxes .== 0) == 3
         print("WARNING: at least one axis should be set to 1 e.g. stripAxes = (1,0,0), otherwise no stripes will be added\n")
@@ -106,20 +106,20 @@ function add_stripes!(Phase, Grid::AbstractGeneralGrid;                # require
 
     Rot3D!(Xrot,Yrot,Zrot, StrikeAngle, DipAngle)
 
-    ph_ind  = findall(Phase .== phase.phase);
+    ph_ind  = findall(Phase .== phase.phase)
 
     ind = Int64[]
     if stripAxes[1] == 1
-        indX     = findall( abs.(Xrot[ph_ind] .% stripeSpacing) .<= stripeWidth/2.0);
-        ind = vcat(ind,indX);
+        indX = findall( abs.(Xrot[ph_ind] .% stripeSpacing) .≤ stripeWidth / 2);
+        ind  = vcat(ind,indX);
     end
     if stripAxes[2] == 1
-        indY     = findall( abs.(Yrot[ph_ind] .% stripeSpacing) .<= stripeWidth/2.0);
-        ind = vcat(ind,indY);
+        indY = findall( abs.(Yrot[ph_ind] .% stripeSpacing) .≤ stripeWidth / 2);
+        ind  = vcat(ind,indY);
     end
     if stripAxes[3] == 1
-        indZ     = findall( abs.(Zrot[ph_ind] .% stripeSpacing) .<= stripeWidth/2.0);
-        ind = vcat(ind,indZ);
+        indZ = findall( abs.(Zrot[ph_ind] .% stripeSpacing) .≤ stripeWidth / 2);
+        ind  = vcat(ind,indZ);
     end
 
     Phase[ph_ind[ind]] .= stripePhase.phase;
@@ -227,9 +227,9 @@ function add_box!(Phase, Temp, Grid::AbstractGeneralGrid;       # required input
     # Set phase number & thermal structure in the full domain
     ztop = maximum(zlim) - Origin[3]
     zbot = minimum(zlim) - Origin[3]
-    ind = findall(  (Xrot .>= (minimum(xlim) - Origin[1])) .& (Xrot .<= (maximum(xlim) - Origin[1])) .&
-                    (Yrot .>= (minimum(ylim) - Origin[2])) .& (Yrot .<= (maximum(ylim) - Origin[2])) .&
-                    (Zrot .>= zbot) .& (Zrot .<= ztop)  )
+    ind  = ((maximum(xlim) - Origin[1]) .≥ Xrot .≥ (minimum(xlim) - Origin[1])) .&
+           ((maximum(ylim) - Origin[2]) .≥ Yrot .≥ (minimum(ylim) - Origin[2])) .&
+           (ztop .≥ Zrot .≥ zbot)
 
     # Compute thermal structure accordingly. See routines below for different options
     if T != nothing 
@@ -311,26 +311,24 @@ function add_layer!(Phase, Temp, Grid::AbstractGeneralGrid;     # required input
     X,Y,Z = coordinate_grids(Grid, cell=cell)
 
     # Limits of block
-    if isnothing(xlim)==isnothing(ylim)==isnothing(zlim)
+    if isnothing(xlim) && isnothing(ylim) && isnothing(zlim)
[…]
