+++ Project.toml
@@ -55,7 +55,7 @@ Glob = "1.2 - 1.3"
 GridapGmsh = "0.5 - 0.7"
 ImageIO = "0.1 - 0.6"
 Interpolations = "0.14, 0.15"
-JLD2 = "0.4"
+JLD2 = "0.4, 0.5"
 LightXML = "0.8, 0.9"
 MeshIO = "0.1 - 0.4"
 NearestNeighbors = "0.2 - 0.4"
