### 2024-09-20 4e3d57d
update test; bugfix for reading
