### 2024-09-20 91d9c58
directly read back NamedTuple
