### 2024-09-20 df320ef
add ASAGI test
