+++ test/test_ASAGI_IO.jl
@@ -1,4 +1,4 @@
-using GeophysicalModelGenerator, Test
+using GeophysicalModelGenerator, Test, Statistics
 
 XYZ                 =   xyz_grid(1.0:1:10.0, 11.0:1:21.0, -23:1:-10);
 Dat                 =   zeros(size(XYZ[1]));
@@ -13,3 +13,13 @@ fname_asagi = write_ASAGI("test", Data)
 # Read back file:
 Data_ASAGI = read_ASAGI(fname_asagi)
 @test sum(Data_ASAGI.fields.Sxx - Data.fields.Sxx) == 0
+@test eltype(Data_ASAGI.fields.Rho[10]) == Float64
+
+# Read back SeisSol file:
+Data_SeisSol = read_ASAGI("test_files/tpv34_rhomulambda-inner.nc")
+@test mean(Data_SeisSol.fields.rho) ≈ 2635.4805f0
+@test eltype(Data_SeisSol.fields.rho[10]) == Float32
+
+
+# Cleanup
+foreach(rm, filter(endswith(".nc"), readdir()))
