### 2024-09-20 a12fcf7
check conversion to km
