### 2024-09-20 fa2496d
removefield for Symbol name
