### 2024-09-20 4118b22
add tests for removefield; allow specifying an Tuple of fields to be removed
