### 2024-09-24 #6763 @Alexander-Barth
@rwegener2, as you probably already know, Boris Kaus was reviewer of my paper (https://joss.theoj.org/papers/10.21105/joss.06504), but this should not affect my ability to "impartial scientific judgment or evaluation" as I see it.
