+++ .github/workflows/SpellCheck.yml
@@ -10,6 +10,6 @@ jobs:
       - name: Checkout Actions Repository
         uses: actions/checkout@v4
       - name: Check spelling
-        uses: crate-ci/typos@v1.23.6
+        uses: crate-ci/typos@v1.25.0
         with:
             args: --exclude **/*.txt --exclude **/*.pvsm
