### 2024-10-08 53275a7
add Chmy interface and multiple dispatch for most functions
