+++ Project.toml
@@ -1,34 +1,7 @@
-authors = ["Boris Kaus", "Marcel Thielmann"]
 name = "GeophysicalModelGenerator"
 uuid = "3700c31b-fa53-48a6-808a-ef22d5a84742"
+authors = ["Boris Kaus", "Marcel Thielmann"]
 version = "0.7.8"
-
-[compat]
-Chmy = "0.1.20"
-Colors = "0.9 - 0.12"
-Downloads = "1"
-FFMPEG = "0.4"
-FileIO = "1"
-GDAL_jll = "300.900.0 - 301.901.0"
-GLMakie = "0.8, 0.9, 0.10"
-GMT = "1.0 - 1.14"
-GeoParams = "0.2 - 0.6"
-Geodesy = "1"
-GeometryBasics = "0.1 - 0.4"
-Glob = "1.2 - 1.3"
-GridapGmsh = "0.5 - 0.7"
-ImageIO = "0.1 - 0.6"
-Interpolations = "0.14, 0.15"
-JLD2 = "0.4, 0.5"
-LightXML = "0.8, 0.9"
-MeshIO = "0.1 - 0.4"
-NearestNeighbors = "0.2 - 0.4"
-Parameters = "0.9 - 0.12"
-SpecialFunctions = "1.0, 2"
-StaticArrays = "1"
-WhereTheWaterFlows = "0.10, 0.11"
-WriteVTK = "1"
-julia = "1.9"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
@@ -58,23 +31,52 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 WhereTheWaterFlows = "ea906314-1493-4d22-a0af-f886a20c9fba"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
+[weakdeps]
+Chmy = "33a72cf0-4690-46d7-b987-06506c2248b9"
+GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+GMT = "5752ebe1-31b9-557e-87aa-f909b540aa54"
+GridapGmsh = "3025c34a-b394-11e9-2a55-3fee550c04c8"
+
 [extensions]
+Chmy_utils = "Chmy"
 GLMakie_Visualisation = "GLMakie"
 GMT_utils = "GMT"
 Gmsh_utils = "GridapGmsh"
-Chmy_utils = "Chmy"
+
+[compat]
+Chmy = "0.1.20"
+Colors = "0.9 - 0.12"
+Downloads = "1"
+FFMPEG = "0.4"
+FileIO = "1"
+GDAL_jll = "300.900.0 - 301.901.0"
+GLMakie = "0.8, 0.9, 0.10"
+GMT = "1.0 - 1.14"
+GeoParams = "0.2 - 0.6"
+Geodesy = "1"
+GeometryBasics = "0.1 - 0.4"
+Glob = "1.2 - 1.3"
+GridapGmsh = "0.5 - 0.7"
+ImageIO = "0.1 - 0.6"
+Interpolations = "0.14, 0.15"
+JLD2 = "0.4, 0.5"
+LightXML = "0.8, 0.9"
+MeshIO = "0.1 - 0.4"
+NearestNeighbors = "0.2 - 0.4"
+Parameters = "0.9 - 0.12"
+SpecialFunctions = "1.0, 2"
+StaticArrays = "1"
+WhereTheWaterFlows = "0.10, 0.11"
+WriteVTK = "1"
+julia = "1.9"
 
 [extras]
+Chmy = "33a72cf0-4690-46d7-b987-06506c2248b9"
[…]
