### 2024-10-09 0cf7795
add functionality for above/belowSurface
