+++ .typos.toml
@@ -2,6 +2,8 @@
 MOR = "MOR"
 dum = "dum"
 Shepard = "Shepard"
+arange = "arange"
+iy = "iy"
 
 [files]
 extend-exclude = ["tutorials/*.pvsm"]
