### 2024-10-09 7b4c24e
changes to make this work with purely 2D arrays
