+++ src/data_types.jl
@@ -1089,14 +1089,14 @@ end
     Structure that holds data for an orthogonal cartesian grid, which can be described with 1D vectors
 """
 struct CartGrid{FT, D} <: AbstractGeneralGrid
-    ConstantΔ   :: Bool                             # Constant spacing (true in all cases for now)
-    N           :: NTuple{D,Int}                    # Number of grid points in every direction
-    Δ           :: NTuple{D,FT}                     # (constant) spacing in every direction
-    L           :: NTuple{D,FT}                     # Domain size
-    min         :: NTuple{D,FT}                     # start of the grid in every direction
-    max         :: NTuple{D,FT}                     # end of the grid in every direction
-    coord1D     :: NTuple{D,AbstractVector{FT}}     # Tuple with 1D vectors in all directions
-    coord1D_cen :: NTuple{D,AbstractVector{FT}}     # Tuple with 1D vectors of center points in all directions
+    ConstantΔ   :: Bool                         # Constant spacing (true in all cases for now)
+    N           :: NTuple{D,Int}                # Number of grid points in every direction
+    Δ           :: NTuple{D,FT}                 # (constant) spacing in every direction
+    L           :: NTuple{D,FT}                 # Domain size
+    min         :: NTuple{D,FT}                 # start of the grid in every direction
+    max         :: NTuple{D,FT}                 # end of the grid in every direction
+    coord1D     :: NTuple{D,Vector{FT}}   # Tuple with 1D vectors in all directions
+    coord1D_cen :: NTuple{D,Vector{FT}}   # Tuple with 1D vectors of center points in all directions
 end
 size(d::CartGrid) = d.N
 
@@ -1253,11 +1253,20 @@ end
 Returns 3D coordinate arrays
 """
 function coordinate_grids(Data::CartGrid; cell=false)
-    X,Y,Z = xyz_grid(NumValue(Data.coord1D[1]), NumValue(Data.coord1D[2]), NumValue(Data.coord1D[3]))
+
+    x_vec = NumValue(Data.coord1D[1])
+    y_vec = NumValue(Data.coord1D[2])
+    z_vec = NumValue(Data.coord1D[3])
 
     if cell
-        X,Y,Z = average_q1(X),average_q1(Y), average_q1(Z)
+        x_vec = (x_vec[2:end] + x_vec[1:end-1])/2
+        z_vec = (z_vec[2:end] + z_vec[1:end-1])/2
+        if length(y_vec)>1
+            y_vec = (y_vec[2:end] + y_vec[1:end-1])/2
+        end
     end
+    
+    X,Y,Z = xyz_grid(x_vec, y_vec, z_vec)
 
     return X,Y,Z
 end
