### 2024-10-09 48b6791
deactivate test that fails locally but runs on CI
