### 2024-10-09 95f4335
bugfix for add_sphere!, more tests
