+++ test/test_lamem.jl
@@ -111,6 +111,10 @@ add_cylinder!(Phases,Temp,Grid, base=(0,0,-5), cap=(3,3,-2), radius=1.5, phase=C
 @test Temp[55,46,45]   == 400.0
 @test Temp[55,45,45]   == 800.0
 
+# for debugging:
+data = CartData(Grid, (;Phases, Temp));
+write_paraview(data,"test")
+
 # test adding generic volcano topography
 Grid = read_LaMEM_inputfile("test_files/SaltModels.dat");
 Topo = make_volc_topo(Grid, center=[0.0,0.0], height=0.4, radius=1.5, crater=0.5, base=0.1);
