+++ test/runtests.jl
@@ -2,6 +2,7 @@ using GeophysicalModelGenerator
 using Test
 
 @testset verbose = true "GeophysicalModelGenerator" begin
+    
     @testset "Data import.jl" begin
         include("test_data_import.jl")
     end
@@ -80,6 +81,7 @@ using Test
     @testset "ASAGI_IO" begin
         include("test_ASAGI_IO.jl")
     end
+    
     @testset "Chmy" begin
         include("test_Chmy.jl")
     end
