### 2024-10-09 231db49
add released 1.11 to CI
