+++ .github/workflows/CI.yml
@@ -16,7 +16,7 @@ jobs:
         version:
           - '1.9'
           - '1.10'
-          - '~1.11.0-0'
+          - '1.11'
           - 'nightly'
         os:
           - ubuntu-latest
