+++ .github/workflows/CI.yml
@@ -21,6 +21,7 @@ jobs:
         os:
           - ubuntu-latest
           - macOS-latest
+          - macos-14
           - windows-latest
         arch:
           - x64
