### 2024-10-09 4a1a436
Update src/Setup_geometry.jl
Co-authored-by: Albert de Montserrat <58044444+albert-de-montserrat@users.noreply.github.com>
