+++ src/Setup_geometry.jl
@@ -664,7 +664,7 @@ julia> write_paraview(Model3D,"LaMEM_ModelSetup")           # Save model to para
 
 """
 function add_polygon!(Phase, Temp, Grid::AbstractGeneralGrid;   # required input
-    xlim=(), ylim::Tuple = (0,0.8), zlim=(),           # limits of the box
+    xlim=(), ylim::Tuple = (0.0,0.8), zlim=(),           # limits of the box
     phase = ConstantPhase(1),                                   # Sets the phase number(s) in the box
     T=nothing, cell=false )                             # Sets the thermal structure (various functions are available)
 
