### 2024-10-10 b0ff3cf
fix bug in document; add missing compat
