+++ tutorials/Tutorial_Chmy_MPI.jl
@@ -125,7 +125,9 @@ MPI.Finalize()
 # such that `mpiexecjl` is created on the command line.
 #
 # You can than run it with:
+# ```
 # mpiexecjl -n 4 --project=. julia Tutorial_Chmy_MPI.jl
+# ```
 
 # The full file can be downloaded [here](../../../tutorials/Tutorial_Chmy_MPI.jl)
 
