### 2024-10-10 610b57b
more tutorial corrections
