### 2024-10-18 #6763 @Alexander-Barth
@rwegener2 Thank you for the reminder :-)

The paper and source code are exceptionally well documented. It is written with new users in mind, which makes it very useful for a large audience. The authors mention that this repository have been used during courses and lectures. Indeed, the authors’ pedagogical approach is clearly seen in this submission.

I have only minor suggestions, and potential typos.

* Would it be useful to show also the in the "1. Getting started" a screenshot with the necessary settings in ParaView (as it is done section 2.)?
https://juliageodynamics.github.io/GeophysicalModelGenerator.jl/dev/man/Tutorial_Basic/

* I am following the documentation, I am wondering if the -3% should not be a 3%
https://github.com/JuliaGeodynamics/GeophysicalModelGenerator.jl/blob/main/docs/src/man/tutorial_load3DSeismicData.md?plain=1#L167
(It seems to the 3% in the screenshot, and I got a similar plot only when using 3% for the iso-surfaces plot)

* I guess read_LaMEM_inputfile should be  `read_LaMEM_inputfile` in the documentation with the added backticks, otherwise LaMEM is in cursive font.
(https://juliageodynamics.github.io/GeophysicalModelGenerator.jl/dev/man/listfunctions/, https://github.com/JuliaGeodynamics/GeophysicalModelGenerator.jl/blob/800d184a4bde6f6b112f5529e9af1bdbb7bbff1d/src/Setup_geometry.jl#L537

* I am wondering if this should that be a bullet list (with `*` to preserve the formatting)
https://github.com/JuliaGeodynamics/GeophysicalModelGenerator.jl/blob/ecddbdfc706ad035587b0b3a4cf8a224544ae377/src/Setup_geometry.jl#L1486-L1499
