### 2024-10-18 011e2b0
address some of the review comments
