+++ src/Setup_geometry.jl
@@ -16,7 +16,7 @@ export  add_box!, add_sphere!, add_ellipsoid!, add_cylinder!, add_layer!, add_po
         ConstantTemp, LinearTemp, HalfspaceCoolingTemp, SpreadingRateTemp, LithosphericTemp, LinearWeightedTemperature,
         McKenzie_subducting_slab,
         ConstantPhase, LithosphericPhases,
-        Trench,
+        Trench, compute_slab_surface
         compute_thermal_structure, compute_phase
 
 """
@@ -71,7 +71,7 @@ Example
 ========
 
 Example: Box with striped phase and constant temperature & a dip angle of 10 degrees:
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 LaMEM Grid:
   nel         : (32, 32, 32)
@@ -175,7 +175,7 @@ Examples
 ========
 
 Example 1) Box with constant phase and temperature & a dip angle of 10 degrees:
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 LaMEM Grid:
   nel         : (32, 32, 32)
@@ -194,7 +194,7 @@ julia> write_paraview(Model3D,"LaMEM_ModelSetup")           # Save model to para
 ```
 
 Example 2) Box with halfspace cooling profile
-```julia
+```julia-repl
 julia> Grid = CartData(xyz_grid(-1000:10:1000,0,-660:10:0))
 julia> Phases = zeros(Int32,   size(Grid));
 julia> Temp   = zeros(Float64, size(Grid));
@@ -291,7 +291,7 @@ Examples
 ========
 
 Example 1) Layer with constant phase and temperature
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 LaMEM Grid:
   nel         : (32, 32, 32)
@@ -310,7 +310,7 @@ julia> write_paraview(Model3D,"LaMEM_ModelSetup")           # Save model to para
 ```
 
 Example 2) Box with halfspace cooling profile
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 julia> Phases = zeros(Int32,   size(Grid.X));
 julia> Temp   = zeros(Float64, size(Grid.X));
@@ -393,7 +393,7 @@ Example
 ========
 
 Sphere with constant phase and temperature:
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 LaMEM Grid:
   nel         : (32, 32, 32)
@@ -462,7 +462,7 @@ Example
 ========
 
 Ellipsoid with constant phase and temperature, rotated 90 degrees and tilted by 45 degrees:
-```julia
+```julia-repl
 julia> Grid = read_LaMEM_inputfile("test_files/SaltModels.dat")
 LaMEM Grid:
   nel         : (32, 32, 32)
@@ -534,7 +534,7 @@ Parameters
 ====
 - `Phase` - Phase array (consistent with Grid)
 - `Temp`  - Temperature array (consistent with Grid)
-- `Grid` - Grid structure (usually obtained with read_LaMEM_inputfile)
+- `Grid` - Grid structure (usually obtained with `read_LaMEM_inputfile`)
 - `base` - center coordinate of bottom of cylinder
 - `cap` - center coordinate of top of cylinder
 - `radius` - radius of the cylinder
@@ -547,7 +547,7 @@ Example
 ========
 
 Cylinder with constant phase and temperature:
-```julia
+```julia-repl
[…]
