+++ tutorials/Tutorial_Basic.jl
@@ -21,9 +21,17 @@ Topo_Alps = load_GMG("https://zenodo.org/records/10738510/files/AlpsTopo.jld2?do
 # We can write this to disk as well
 write_paraview(Topo_Alps,"Topo_Alps")
 
-# If we open both datasets in Paraview, we see this (after giving some color to the topography):
+# If we open both datasets in Paraview, and changing both files from outline/solid colors to the corresponing data field, we see:
+# ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_Paraview_1.png)
+# Now we can change the colormap on the right side, marked by a red square. For topography we use the `Oleron` colormap, which you can download [here](https://www.fabiocrameri.ch/colourmaps/).
+# For the tomography we use the `Roma` scientific colormap. You will now see a blue'ish box of the tomography, this is not the best color to visualise the data. Let's invert the colormap by clicking on the item marked by the blue arrow.
+# Now we see the tomography in a more intuitive way, but the topography is not visible anymore. We can change the opacity of the tomography by setting a value in the `Opacity` field marked by the red square.
+# Note that you will need to adapt the range of the topography colormap as the change in color is not at 0.0. By clicking on the item marked by the black arrow, you can set your desired range.
+
+# ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_Paraview_2.png)
+
+# Now you should see somthing like this:
 # ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_1.png)
-# Note that I use the `Oleron` scientific colormap for the tomography which you can download [here](https://www.fabiocrameri.ch/colourmaps/)
 #
 # ### 2. Extract subset of data
 # As you can see the tomographic data covers a much larger area than the Alps itself, and in most of that area there is no data.
@@ -32,7 +40,7 @@ Tomo_Alps = extract_subvolume(Tomo_Alps_full,Lon_level=(4,20),Lat_level=(36,50),
 
 write_paraview(Tomo_Alps,"Tomo_Alps");
 
-# Which looks like:
+# After loading the new data again in paraview, switching to the proper data field and adjusting the colormap, you should see somthing like this:
 # ![Basic_Tutorial_2](../assets/img/Basic_Tutorial_2.png)
 
 # ### 3. Create cross sections
@@ -45,40 +53,40 @@ data_200km = cross_section(Tomo_Alps, Depth_level=-200)
 data_200km_exact = cross_section(Tomo_Alps, Depth_level=-200, Interpolate=true)
 
 # In general, you can get help info for all functions with `?`:
-# ```julia
+# ```julia-repl
 # help?> cross_section
 # search: cross_section cross_section_volume cross_section_points cross_section_surface flatten_cross_section
-# 
+#
 #   cross_section(DataSet::AbstractGeneralGrid; dims=(100,100), Interpolate=false, Depth_level=nothing, Lat_level=nothing, Lon_level=nothing, Start=nothing, End=nothing, Depth_extent=nothing, section_width=50km)
-# 
+#
 #   Creates a cross-section through a GeoData object.
-# 
+#
 #     •  Cross-sections can be horizontal (map view at a given depth), if Depth_level is specified
-# 
+#
 #     •  They can also be vertical, either by specifying Lon_level or Lat_level (for a fixed lon/lat), or by defining both Start=(lon,lat) & End=(lon,lat) points.
-# 
+#
[…]
