+++ docs/src/man/Tutorial_Basic.md
@@ -21,7 +21,7 @@ Tomo_Alps_full = load_GMG("https://zenodo.org/records/10738510/files/Paffrath_20
 ```
 
 ````
-GeoData 
+GeoData
   size      : (162, 130, 42)
   lon       ϵ [ -13.3019 : 35.3019]
   lat       ϵ [ 30.7638 : 61.2362]
@@ -49,7 +49,7 @@ Topo_Alps = load_GMG("https://zenodo.org/records/10738510/files/AlpsTopo.jld2?do
 ```
 
 ````
-GeoData 
+GeoData
   size      : (961, 841, 1)
   lon       ϵ [ 4.0 : 20.0]
   lat       ϵ [ 36.0 : 50.0]
@@ -70,7 +70,7 @@ Saved file: Topo_Alps.vts
 
 ````
 
-If we open both datasets in Paraview, and changing both files from outline/solid colors to the corresponing data field, we see:
+If we open both datasets in Paraview, and changing both files from outline/solid colors to the corresponding data field, we see:
 ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_Paraview_1.png)
 Now we can change the colormap on the right side, marked by a red square. For topography we use the `Oleron` colormap, which you can download [here](https://www.fabiocrameri.ch/colourmaps/).
 For the tomography we use the `Roma` scientific colormap. You will now see a blue'ish box of the tomography, this is not the best color to visualise the data. Let's invert the colormap by clicking on the item marked by the blue arrow.
@@ -79,7 +79,7 @@ Note that you will need to adapt the range of the topography colormap as the cha
 
 ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_Paraview_2.png)
 
-Now you should see somthing like this:
+Now you should see something like this:
 ![Basic_Tutorial_1](../assets/img/Basic_Tutorial_1.png)
 
 ### 2. Extract subset of data
@@ -97,7 +97,7 @@ Saved file: Tomo_Alps.vts
 
 ````
 
-After loading the new data again in paraview, switching to the proper data field and adjusting the colormap, you should see somthing like this:
+After loading the new data again in paraview, switching to the proper data field and adjusting the colormap, you should see something like this:
 ![Basic_Tutorial_2](../assets/img/Basic_Tutorial_2.png)
 
 ### 3. Create cross sections
@@ -110,7 +110,7 @@ data_200km = cross_section(Tomo_Alps, Depth_level=-200)
 ```
 
 ````
-GeoData 
+GeoData
   size      : (54, 60, 1)
   lon       ϵ [ 3.9057 : 19.9057]
   lat       ϵ [ 35.9606 : 49.8976]
@@ -126,7 +126,7 @@ data_200km_exact = cross_section(Tomo_Alps, Depth_level=-200, Interpolate=true)
 ```
 
 ````
-GeoData 
+GeoData
   size      : (100, 100, 1)
   lon       ϵ [ 3.9057 : 19.9057]
   lat       ϵ [ 35.9606 : 49.8976]
@@ -184,7 +184,7 @@ Cross_vert = cross_section(Tomo_Alps, Start=(5,47), End=(15,44))
 ```
 
 ````
-GeoData 
+GeoData
   size      : (100, 100, 1)
   lon       ϵ [ 5.0 : 15.0]
   lat       ϵ [ 47.0 : 44.0]
@@ -225,7 +225,7 @@ Topo_cart = convert2CartData(Topo_Alps, proj)
 ```
 
 ````
-CartData 
+CartData
     size    : (961, 841, 1)
[…]
