+++ README.md
@@ -45,12 +45,12 @@ First, you need to install julia on your machine. We recommend to use the binari
 Next, start julia and switch to the julia package manager using `]`, after which you can add the package.
 ```julia-repl
 julia> ]
-(@1.11) pkg> add GeophysicalModelGenerator
+(@1.6) pkg> add GeophysicalModelGenerator
 ```
 You can test whether it works on your system with
 ```julia-repl
 julia> ]
-(@1.11) pkg> test GeophysicalModelGenerator
+(@1.6) pkg> test GeophysicalModelGenerator
 ```
 and use it with
 ```julia-repl
