### 2024-10-18 #6763 @boriskaus
Hi @Alexander-Barth
Thanks for your suggestions.  They have now been all implemented in the `GeophysicalModelGenerator` and are part of version 0.7.10
