### 2024-10-25 #6763 @rwegener2
## 1 - Editor Tasks Prior to Acceptance
Editor checks paper proof:
- [ ] Download the proof and check rendering
- [ ] Check all references have DOIs, follow the links and check the references with ``@editorialbot check references`` and ask author(s) to update as needed

## 3 - Editor Archiving and Submission Recommendation
Editor checks archive generated by author:
- [ ] Check that the archive title, author list, and version tag are the same between the archive and the paper
- [ ] Check the license is the same between the archive, the paper, and the Github repository

Editor double checks paper and recommends submission:
