### 2024-10-25 #6763 @rwegener2
> @rwegener2 what are the next steps?

@boriskaus I just created a checklist above which outline the process of my final checks. I will complete the first set of steps, hand it back to you, complete my final steps, then move the submission back to the track editor. The track editor's final checks are the last steps before publication.
