### 2024-10-25 #6763 @rwegener2
Hello @boriskaus! Nice article. A few small comments (within any quotes from your paper ~strikethrough~ is a possible word to be deleted and **bold** is a possible word to be added):

1. Would you please look into the citation "Kaus, B. J., Popov, A. A., Baumann, T., Pusok, A., Bauville, A., Fernandez, N., & Collignon, M. (2016). Forward and inverse modelling of lithospheric deformation on geological timescales. NIC Series, 48, 299–306. https://doi.org/2128/9842"? I currently get an error when following the doi link https://doi.org/2128/9842
2. The sentence that starts in line 116 seems to be either missing a word or have an extra word. Did you mean "It is also possible to interpolate other seismic tomography datasets to the same grid and subsequently compute a votemap to count ~in~ how many tomographic models **contian** a specific seismic anomaly ~is present~"?
