### 2024-10-25 #6763 @boriskaus
1. The publication is an extended abstract that is available here: https://juser.fz-juelich.de/record/507751/files/nic_2016_kaus.pdf it doesn’t seem to have a DOI, perhaps because it was part of a conference proceeding book. How would we best cite that instead?
2. Your suggestion is fine, when we use “contain” instead of “contian”
