+++ docs/paper/paper.bib
@@ -88,7 +88,8 @@ DOI = {10.5194/se-10-1785-2019}
  booktitle={NIC Series}, 
  author={Kaus, Boris JP and Popov, Anton A. and Baumann, T. and Pusok, A. and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine}, 
  year={2016}, 
- pages={299-306} }
+ pages={299-306}
+ note={\url{http://hdl.handle.net/2128/9842}} }
 
 @article{napolitano2023imaging,
   title={Imaging overpressurised fracture networks and geological barriers hindering fluid migrations across a slow-deformation seismic gap},
