+++ docs/paper/paper.bib
@@ -86,7 +86,7 @@ DOI = {10.5194/se-10-1785-2019}
  ISBN={978-3-95806-109-5}, 
  url={http://hdl.handle.net/2128/9842}, 
  booktitle={NIC Series}, 
- author={Kaus, Boris JP and Popov, Anton A. and Baumann, T. and Pusok, A. and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine}, 
+ author={Kaus, Boris JP and Popov, Anton A. and Baumann, T. and Püsök, A. and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine}, 
  year={2016}, 
  pages={299-306},
  note={\url{http://hdl.handle.net/2128/9842}} }
