### 2024-10-25 60bacf4
direct link to book chapter rather than full book
