+++ docs/paper/paper.bib
@@ -89,7 +89,7 @@ DOI = {10.5194/se-10-1785-2019}
  author={Kaus, Boris JP and Popov, Anton A. and Baumann, T. and Püsök, A. and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine}, 
  year={2016}, 
  pages={299-306},
- note={\url{http://hdl.handle.net/2128/9842}} }
+ note={\url{https://juser.fz-juelich.de/record/507751/files/nic_2016_kaus.pdf}} }
 
 @article{napolitano2023imaging,
   title={Imaging overpressurised fracture networks and geological barriers hindering fluid migrations across a slow-deformation seismic gap},
