+++ docs/paper/paper.bib
@@ -84,13 +84,12 @@ DOI = {10.5194/se-10-1785-2019}
  volume={48}, 
  rights={All rights reserved}, 
  ISBN={978-3-95806-109-5}, 
- url={http://hdl.handle.net/2128/9842}, 
+ url={https://juser.fz-juelich.de/record/507751/files/nic_2016_kaus.pdf}, 
  booktitle={NIC Series}, 
  author={Kaus, Boris JP and Popov, Anton A. and Baumann, T. and Püsök, A. and Bauville, Arthur and Fernandez, Naiara and Collignon, Marine}, 
  year={2016}, 
- pages={299-306},
- note={\url{https://juser.fz-juelich.de/record/507751/files/nic_2016_kaus.pdf}} }
-
+ pages={299-306}
+}
 @article{napolitano2023imaging,
   title={Imaging overpressurised fracture networks and geological barriers hindering fluid migrations across a slow-deformation seismic gap},
   author={Napolitano, Ferdinando and Gabrielli, Simona and De Siena, Luca and Amoroso, Ortensia and Capuano, Paolo},
