### 2024-10-25 #6763 @boriskaus
As to step 2:

> Double check authors and affiliations (including ORCIDs)

done

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

version 0.7.10

> Archive the release on Zenodo/figshare/etc and post the DOI here.
https://zenodo.org/records/13952117

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
done

> Make sure that the license listed for the archive is the same as the software license.
done
