### 2024-10-25 98cd359
bugfix apparently introduced in PR122
