+++ ext/GMT_utils.jl
@@ -68,8 +68,8 @@ julia> write_paraview(Topo,"Topo_Alps")
 function import_topo(limits; file::String="@earth_relief_01m", maxattempts=5)
 
   # Correct if negative values are given (longitude coordinates that are west)
-  ind = limits[1:2] .< 0
-    
+  ind = findall(limits[1:2] .< 0); 
+
   if (limits[1] < 0) && (limits[2] < 0)
       limits[ind] .= 360 .+ limits[ind]
       limits[1:2]  = sort(limits[1:2])   
