+++ Project.toml
@@ -51,7 +51,7 @@ FFMPEG = "0.4"
 FileIO = "1"
 GDAL_jll = "300.900.0 - 301.901.0"
 GLMakie = "0.8, 0.9, 0.10"
-GMT = "1.0 - 1.20"
+GMT = "1.0 - 1.14"
 GeoParams = "0.2 - 0.6"
 Geodesy = "1"
 GeometryBasics = "0.1 - 0.4"
