+++ Project.toml
@@ -10,7 +10,6 @@ Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FFMPEG = "c87230d0-a227-11e9-1b43-d7ebe4e7570a"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 GDAL_jll = "a7073274-a066-55f0-b90d-d619367d196c"
-GMT = "5752ebe1-31b9-557e-87aa-f909b540aa54"
 GeoParams = "e018b62d-d9de-4a26-8697-af89c310ae38"
 Geodesy = "0ef565a4-170c-5f04-8de2-149903a85f3d"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
@@ -35,13 +34,14 @@ WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 [weakdeps]
 Chmy = "33a72cf0-4690-46d7-b987-06506c2248b9"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+GMT = "5752ebe1-31b9-557e-87aa-f909b540aa54"
 GridapGmsh = "3025c34a-b394-11e9-2a55-3fee550c04c8"
 
 [extensions]
 Chmy_utils = "Chmy"
 GLMakie_Visualisation = "GLMakie"
-GMT_utils = "GMT"
 Gmsh_utils = "GridapGmsh"
+GMT_utils = "GMT"
 
 [compat]
 Chmy = "0.1.20"
@@ -51,11 +51,11 @@ FFMPEG = "0.4"
 FileIO = "1"
 GDAL_jll = "300.900.0 - 301.901.0"
 GLMakie = "0.8, 0.9, 0.10"
-GMT = "1.0 - 1.14"
 GeoParams = "0.2 - 0.6"
 Geodesy = "1"
 GeometryBasics = "0.1 - 0.4"
 Glob = "1.2 - 1.3"
+GMT = "1"
 GridapGmsh = "0.5 - 0.7"
 ImageIO = "0.1 - 0.6"
 Interpolations = "0.14, 0.15"
@@ -74,7 +74,6 @@ julia = "1.9"
 [extras]
 Chmy = "33a72cf0-4690-46d7-b987-06506c2248b9"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
-GMT = "5752ebe1-31b9-557e-87aa-f909b540aa54"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
