### 2024-10-30 #6763 @rwegener2
Hi @boriskaus. Typically in JOSS submissions the zenodo archive has a title which matches the paper title. It looks like you have used a pre-existing zenodo archive which has a different title from the paper.

Do you want to move forward with the submission using the existing archive or create a new archive with a matching title? It's fine to move forward with the existing archive, but I wanted to offer the option to change if desired.
