### 2024-10-30 #6763 @boriskaus
Hi @rwegener2, I am using the option of GitHub to automatically generate a zenodo repository every time a new release is made.

Ideally, I would like to keep using that, which indeed gives a slightly different title. In my opinion it is pretty clear that this is about the same, given that it starts with the package name.

So if it’s fine with you I would like to go ahead with what we have now
