### 2024-10-31 #6763 @rwegener2
> Ideally, I would like to keep using that, which indeed gives a slightly different title. In my opinion it is pretty clear that this is about the same, given that it starts with the package name.

> So if it’s fine with you I would like to go ahead with what we have now

Sounds good, @boriskaus!
