### 2024-10-31 #6763 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
