### 2024-11-04 #6763 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
