+++ docs/paper/paper.bib
@@ -68,7 +68,7 @@ DOI = {10.5194/se-10-1785-2019}
 }
 
 @article{gabrielli2023scattering,
-  title={Scattering attenuation images of the control of thrusts and fluid overpressure on the 2016--2017 Central Italy seismic sequence},
+  title={Scattering attenuation images of the control of thrusts and fluid overpressure on the 2016--2017 {C}entral {I}taly seismic sequence},
   author={Gabrielli, Simona and Akinci, Aybige and De Siena, Luca and Del Pezzo, Edoardo and Buttinelli, Mauro and Maesano, Francesco Emanuele and Maffucci, Roberta},
   journal={Geophysical Research Letters},
   volume={50},
